Improve buildingConnector API
