+++ ShapezShifter/manifest.json
@@ -1,6 +1,6 @@
 {
   "Name": "tobspr.ShapezShifter",
-  "Version": "0.9.0",
+  "Version": "0.9.1",
   "Title": "Shapez 2 - Extended API",
   "Description": "Extended functionality, convenience and mod support for shapez 2",
   "Author": "tobspr Games",
